Fix t0275-cid-security.sh: rename bad block file
This sharness test provides 2 datastore blocks directly.

One of them (corresponding to an "insecure block") is not a CIDv0 therefore it
needs to be migrated and stored with its raw-multihash ID.

The (raw) CID of the block used in the test remains the same.

The other block needs no changes.